+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 - Make it easier to manipulate svg files for Imagick and Gmagick by change file name to jpg
 - Fix a bug on save file when we have multiple instances from a model
 - Bugfix on set LARUPLOAD_NULL
+- store human readable file type into database
 
 ## v0.0.2 - 2018/09/10
 - Support Laravel 5.7